feat: wrapping planets topologically as spheres
Planets are donuts!

It is confusing what kind of grid edge connection it describes.
According to math topology, it describes a torus (donut). **Given a grid from 1 to 5 in x, and 1 to 5 in y** a torus connects (1,1) to (5,1) given x-1 and (1,1) to (1,5) given y-1.
If we want a sphere, we cannot connect one edge of the grid to another. We have to connect only the x edge to another, so (1,1) for x-1 to (5,1), but we have to connect vertical edges towards themselves in inverted coordinates, so (1,1) for y-1 connects to (5,1).

Donut:
|(x,y)  |  x+1  |  x-1  |  y+1  |**y-1**|
|-------|-------|-------|-------|-------|
| (1,1) | (2,1) | (5,1) | (1,2) | (1,5) |
| (2,1) | (3,1) | (1,1) | (2,2) | (2,5) |
| (3,1) | (4,1) | (2,1) | (3,2) | (3,5) |

Sphere:
| (x,y) |  x+1  |  x-1  |  y+1  |**y-1**|
|-------|-------|-------|-------|-------|
| (1,1) | (2,1) | (5,1) | (1,2) | (5,1) |
| (2,1) | (3,1) | (1,1) | (2,2) | (4,1) |
| (3,1) | (4,1) | (2,1) | (3,2) | (3,1) |

More information here: https://www.youtube.com/watch?v=lmcT2mP2bfE